server/requirements.txt
flask
openai-whisper
requests
python-dotenv
setuptools-rust